fix: separate resource code
